Extract modal content into separate component file
